refactor : HomeScreen 코드 가독성 개선
